refactor: move addFunds out of v1
